+++ resources/js/core/Router.vue
@@ -3,7 +3,8 @@
     <transition mode="out-in"
         enter-active-class="fadeIn"
         leave-active-class="fadeOut">
-        <router-view class="animated"/>
+        <router-view class="animated"
+            :key="key"/>
     </transition>
 
 </template>
@@ -12,6 +13,12 @@
 
 export default {
     name: 'Router',
+
+    computed: {
+        key() {
+            return JSON.stringify(this.$route.params);
+        },
+    },
 };
 
 </script>